Serpents Fate V1.0

+++ src/entities/Enemy.java
@@ -79,7 +79,7 @@ Random random = new Random();
     	this.defense = defense;
     }
     
-    public double getOriginalDefense() {
+    public int getOriginalDefense() {
         return originalDefense;
     }
